[FEAT] PostList, useListQuery 공통 컴포넌트로 수정하기(activity 재사용) #NADA-52
