+++ src/containers/community/PostList.js
@@ -1,30 +1,15 @@
 /**PostList 커뮤니티 게시물 리스트가 나오는 컨테이너 */
-import { useSelector, useDispatch } from 'react-redux';
 import { Link } from 'react-router-dom';
-import { changePostDetailField } from '../../modules/community/postDetail';
 import PostItem from '../../components/community/PostList/PostItem';
 import useListQuery from '../../modules/queries/useListQuery';
 import { ErrorMessage } from './ErrorMessage';
-const PostList = () => {
-  const dispatch = useDispatch();
-
-  const filter = useSelector(({ filter }) => filter);
-  const { buttonSelect, subButtonSelect } = filter;
-
+const PostList = ({ type, filter, setData }) => {
   // useQuery를 사용하여 데이터 가져오기
-  const {
-    data: postsData,
-    isLoading,
-    isError,
-  } = useListQuery({
-    mainCategory: buttonSelect.maincategory,
-    region: subButtonSelect.region,
-    field: subButtonSelect.field,
-    category: subButtonSelect.category,
+  const { data, isLoading, isError } = useListQuery({
+    type: type,
+    filter: filter,
   });
-  const setData = (card) => {
-    dispatch(changePostDetailField({ value: card }));
-  };
+
   if (isLoading) {
     return <div>Loading...</div>;
   }
@@ -35,9 +20,9 @@ const PostList = () => {
 
   return (
     <div>
-      {postsData.map((card) => (
+      {data.map((card) => (
         <Link
-          to={`/community/PostDetail/${card._id}`}
+          to={`/${type}/PostDetail/${card._id}`}
           key={card._id}
           onClick={() => setData(card)}
         >
